discord-bot/requirements.txt
discord.py==2.3.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1